+++ jaxley/integrate.py
@@ -57,8 +57,8 @@ def build_init_and_step_fn(
         params = cell.get_parameters()
 
         cell.to_jax()
-        rec_inds = cell.recordings.rec_index.to_numpy()
-        rec_states = cell.recordings.state.to_numpy()
+        rec_inds = cell.rec_info.rec_index.to_numpy()
+        rec_states = cell.rec_info.state.to_numpy()
         externals = cell.externals.copy()
         external_inds = cell.external_inds.copy()